refactor(webpack): colocate manifest stats helpers
