+++ development/webpack/utils/plugins/ManifestPlugin/stats.ts
